Goblin Agent/pyproject.toml
[project]
name = "goblin-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "finnhub-python>=2.4.25",
    "langchain>=1.0.7",
    "langchain-groq>=1.0.1",
    "langgraph>=1.0.3",
    "langgraph-cli[inmem]>=0.4.7",
    "pandas>=2.3.3",
    "ta>=0.11.0",
    "typing>=3.10.0.0",
    "yfinance>=0.2.66",
]
